ClientAdvisor/AzureFunction/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azurefunctions-extensions-http-fastapi==1.0.0b1
openai
semantic_kernel
psycopg2-binary
pgvector